VPalm

Documentation for VPalm.

VPalm.mtg_skeleton

VPalm.read_parameters

VPalm.static_mockup

VPalm.stem_bending

VPalm.stem_height

VPalm.write_parameters

VPalm.mtg_skeleton — Method

mtg_skeleton(nb_leaves_emitted)

Makes an MTG skeleton with nb_leaves_emitted leaves, including all intermediate organs:

Plant: the whole palm

Stem: the stem of the plant, i.e. the remaining part of the plant after the leaves have been removed

Phytomer: the part that includes the leaf and the internode

Internodes: the part of the phytomer that is between two leaves

Leaf: the leaf of the plant, also called frond

Note: this skeleton does not include reproductive organs (inflorescences, fruits) or the scales that decompose the leaf (petiole, rachis, leaflets).

Arguments

nb_leaves_emitted: The number of leaves to emit.

Examples

mtg_skeleton(3)

VPalm.read_parameters — Method

read_parameters(file)

Reads a parameter file and returns the contents as an ordered dictionary.

Arguments

file: The path to the parameter file.

Returns

An ordered dictionary containing the contents of the parameter file.

Example

file = joinpath(dirname(dirname(pathof(VPalm))),"test","files","parameter_file.yml")
read_parameters(file)

VPalm.static_mockup — Method

static_mockup(parameters)

Builds a 3D mockup from a set of parameters. The mockup is static, i.e. it does not change over time.

VPalm.stem_bending — Method

stem_bending(stem_bending_mean, stem_bending_sd; rng=Random.MersenneTwister(1234))

Computes the stem bending (°) using an average bending and a standard deviation (random draw from a normal distribution).

Arguments

stem_bending_mean: The average stem bending (°).

stem_bending_sd: The standard deviation of the stem bending (°).

Optional arguments

rng: The random number generator.

VPalm.stem_height — Method

stem_height(nb_leaves_emitted, initial_stem_height, stem_height_coefficient, internode_length_at_maturity, stem_growth_start)

Computes the stem height (m) at a given number of leaves emitted.

Arguments

nb_leaves_emitted: The number of leaves emitted from planting.

initial_stem_height: The initial stem height at planting (m).

stem_height_coefficient: The coefficient of the exponential function.

internode_length_at_maturity: The internode length when the plant is mature (m).

stem_growth_start: The number of leaves emitted at which the stem starts to grow (m). This is because the stem does not grow at the same rate at the beginning of the plant's life,

because it first grows more in diameter than in height.

stem_height_variation: The variation of the stem height (m) due to the random draw from a normal distribution.

Optional arguments

rng: The random number generator.

Details

The stem height is computed using an exponential function for the first stem_growth_start leaves emitted, and then a linear function for the remaining leaves emitted.

Note that the stem height can also be subject to some variability using stem_height_variation, simulating natural variations that might occur in real-world scenarios, but this variability will never make the stem height go below 30% of the intial computed height.

VPalm.write_parameters — Method

write_parameters(file, params)

Write the given parameters to a file using YAML format.

Arguments

file: The file path to write the parameters to.

params: The parameters to be written.

Example

file = joinpath(dirname(dirname(pathof(VPalm))),"test","files","parameter_file.yml")
params = read_parameters(file)
write_parameters(tempname(), params)