VPalm

Documentation for VPalm.

VPalm.cylinder

VPalm.internode_diameter

VPalm.internode_length

VPalm.leaf_insertion_angle

VPalm.mtg_skeleton

VPalm.phyllotactic_angle

VPalm.read_parameters

VPalm.read_ply

VPalm.snag

VPalm.stem_bending

VPalm.stem_diameter

VPalm.stem_height

VPalm.write_parameters

VPalm.cylinder — Method

cylinder()
cylinder(r, l)

Returns a normalized cylinder mesh, or a cylinder with radius r and length l.

Arguments

r: The radius of the cylinder.

l: The length of the cylinder.

VPalm.internode_diameter — Method

Internode diameter model

Computes the diameter of an internode at a given rank.

Arguments

internode_number: The number of the internode.

nb_internodes: The total number of internodes.

stem_diameter: The diameter of the stem at the base.

stem_base_shrinkage: The shrinkage coefficient at the stem base.

stem_top_shrinkage: The shrinkage coefficient at the stem top.

leaves_in_sheath: The number of leaves in the sheath.

VPalm.internode_length — Method

Internode length model

Computes the length of an internode at a given rank. The internode length is computed using a quadratic function. The objective is to have a internodes that are short and growing for the first emitted leaves (before nb_internodes_before_planting), and then getting to a stable "constant" height, and at the end for the youngest leaves, having nodes currently growing (smaller).

The internode length is computed as follows : Internode length ^ l | _____________________ | /| || / | | l0 | / | | |-|–-|––––––––––-|–-|––> Internode number 1 N N + N + Nbl Nbl - R where : - l0 is internode_min_height (m), the minimum height of the internode. - l is internode_heigth_final (m), the maximum height of the internode.- N isnbinternodesbeforeplanting, the number of internodes before planting. - R isinternoderanknoexpansion, the number of internodes not in expansion. - Nbl is the number of leaves emitted since planting. with the conditions that : - the sum of the areas of the first triangle, the rectangle and the last triangle is equal tostemheight. - if the equation of the first line isa * x + b: -a = (l - l0) / (N - 1)-b = l0 - a- the area of the first triangle isa * N * (N + 1) / 2 + b * Nand after development :l * N/2 + l0 * N/2- the area of the rectangle (between N + 1 and N + Nbl - R - 1) is(Nbl - R - 1) * l- if the equation of the last line isc * x + d, then: -c = (l0 - l) / R-d = l0 - c * (Nbl + N)- the area of the last triangle is(R + 1) * (c * (2N + 2Nbl - R) / 2 + d)and after development :l * ((R + 1)/ 2) + l_0 * (-(R + 1) / 2 + R + 1)reminder: - the sum of integers from m to n isn * (n + 1) / 2 - m * (m - 1) / 2- the sum of cx + d from m to n isc * (n * (n + 1) / 2 - m * (m - 1) / 2) + d * (n - m + 1)or(n - m + 1) * (c * (n + m) / 2 + d)`

Arguments

i/ internode_number: The number of the internode.

Nbl / nb_internodes: The total number of internodes == number of leaves emitted since planting.

sh / stem_height: The height of the stem.

R / internode_rank_no_expansion: The rank of the internode that will not expand.

N / nb_internodes_before_planting: The number of internodes before planting.

l_0 / internode_min_height: The minimal length of the internode.

VPalm.leaf_insertion_angle — Function

leaf_insertion_angle(rank, leaf_max_angle=90, leaf_slope_angle=0.05, leaf_inflection_angle=40)

Compute the insertion angle of the leaf on the internode.

Note: The insertion angle is computed using a logistic function.

Arguments

rank: The rank of the leaf.

leaf_max_angle: The maximum angle of the leaf.

leaf_slope_angle: The slope of the logistic function.

leaf_inflection_angle: The inflection point of the logistic function.

VPalm.mtg_skeleton — Method

mtg_skeleton(nb_internodes)

Makes an MTG skeleton with nb_leaves_emitted leaves, including all intermediate organs:

Plant: the whole palm

Stem: the stem of the plant, i.e. the remaining part of the plant after the leaves have been removed

Phytomer: the part that includes the leaf and the internode

Internodes: the part of the phytomer that is between two leaves

Leaf: the leaf of the plant, also called frond

Note: this skeleton does not include reproductive organs (inflorescences, fruits) or the scales that decompose the leaf (petiole, rachis, leaflets).

Arguments

nb_internodes: The number of internodes to emit.

Examples

file = joinpath(dirname(dirname(pathof(VPalm))), "test", "files", "parameter_file.yml")
parameters = read_parameters(file)
mtg_skeleton(parameters)

VPalm.phyllotactic_angle — Method

phyllotactic_angle(phyllotactic_angle_mean, phyllotactic_angle_sd; rng=Random.MersenneTwister(1234))

Computes the phyllotactic angle (°) using an average angle and a standard deviation (random draw from a normal distribution).

Arguments

phyllotactic_angle_mean: The average phyllotactic angle (°).

phyllotactic_angle_sd: The standard deviation of the phyllotactic angle (°).

Optional arguments

rng: The random number generator.

VPalm.read_parameters — Method

read_parameters(file)

Reads a parameter file and returns the contents as an ordered dictionary.

Arguments

file: The path to the parameter file.

Returns

An ordered dictionary containing the contents of the parameter file.

Example

file = joinpath(dirname(dirname(pathof(VPalm))),"test","files","parameter_file.yml")
read_parameters(file)

VPalm.read_ply — Method

read_ply(fname)

Reads a PLY file and returns a Meshes.SimpleMesh object.

Arguments

fname: The path to the PLY file.

Returns

A Meshes.SimpleMesh object.

VPalm.snag — Method

snag()
snag(l, w, h)

Returns a normalized snag mesh, or a snag mesh with given dimensions in m.

VPalm.stem_bending — Method

stem_bending(stem_bending_mean, stem_bending_sd; rng=Random.MersenneTwister(1234))

Computes the stem bending (°) using an average bending and a standard deviation (random draw from a normal distribution).

Arguments

stem_bending_mean: The average stem bending (°).

stem_bending_sd: The standard deviation of the stem bending (°).

Optional arguments

rng: The random number generator.

VPalm.stem_diameter — Method

stem_diameter(rachis_length_reference, stem_diameter_max, stem_diameter_slope, stem_diameter_inflection, stem_diameter_residual)

Computes the stem diameter (m) at a given rachis length reference (m).

Arguments

rachis_length_reference: The rachis length reference (m). Taken as the rachis length of the first leaf.

stem_diameter_max: The maximum stem diameter (m).

stem_diameter_slope: The slope of the logistic function.

stem_diameter_inflection: The inflection point of the logistic function.

stem_diameter_residual: The residual of the stem diameter (m).

stem_diameter_snag: The diameter estimation due to snags (m).

Optional arguments

rng: The random number generator.

Details

The stem diameter is computed using a logistic function, and then some variability is added to simulate natural variations that might occur in real-world scenarios.

VPalm.stem_height — Method

stem_height(nb_leaves_emitted, initial_stem_height, stem_height_coefficient, internode_length_at_maturity, stem_growth_start)

Computes the stem height (m) at a given number of leaves emitted.

Arguments

nb_leaves_emitted: The number of leaves emitted from planting.

initial_stem_height: The initial stem height at planting (m).

stem_height_coefficient: The coefficient of the exponential function.

internode_length_at_maturity: The internode length when the plant is mature (m).

stem_growth_start: The number of leaves emitted at which the stem starts to grow (m). This is because the stem does not grow at the same rate at the beginning of the plant's life,

because it first grows more in diameter than in height.

stem_height_variation: The variation of the stem height (m) due to the random draw from a normal distribution.

Optional arguments

rng: The random number generator.

Details

The stem height is computed using an exponential function for the first stem_growth_start leaves emitted, and then a linear function for the remaining leaves emitted.

Note that the stem height can also be subject to some variability using stem_height_variation, simulating natural variations that might occur in real-world scenarios, but this variability will never make the stem height go below 30% of the intial computed height.

VPalm.write_parameters — Method

write_parameters(file, params)

Write the given parameters to a file using YAML format.

Arguments

file: The file path to write the parameters to.

params: The parameters to be written.

Example

file = joinpath(dirname(dirname(pathof(VPalm))),"test","files","parameter_file.yml")
params = read_parameters(file)
write_parameters(tempname(), params)